// src/Types/types.tsx
export type issuerData = {
    name: string;
    id: string;
    type: string;
    branchID: string;
    country: string;
    governate: string;
    regionCity: string;
    street: string;
    buildingNumber: string;
    postalCode: string;
    floor: string;
    room: string;
    landmark: string;
    additionalInformation: string;
}